added shape inference test and updated testing strucutre
